Clean up job scheduler, resource manager, and resource consumers
