+++ nvflare/poc/client/startup/fed_client.json
@@ -34,17 +34,17 @@
   "components": [
     {
       "id": "resource_manager",
-      "path": "nvflare.apis.impl.list_resource_manager.ListResourceManager",
+      "name": "ListResourceManager",
       "args":  {
-        "resources": {"gpu": [0,1,2,3]}
-        }
+        "resources": {"gpu": [0, 1, 2, 3]}
+      }
     },
     {
       "id": "resource_consumer",
-      "path": "nvflare.apis.impl.gpu_resource_consumer.GPUResourceConsumer",
+      "name": "GPUResourceConsumer",
       "args":  {
         "gpu_resource_key": "gpu"
-        }
+      }
     }
   ]
 }
